#21 Create Quote(Service)

+++ src/main/java/dev/sorn/fmp4j/models/FmpShortQuote.java
@@ -1,13 +1,12 @@
 package dev.sorn.fmp4j.models;
 
 import java.io.Serial;
-import java.math.BigDecimal;
 
 public record FmpShortQuote(
     String symbol,
-    BigDecimal price,
-    BigDecimal change,
-    BigDecimal volume
+    Double price,
+    Double change,
+    Long volume
 ) implements FmpModel {
     @Serial
     private static final long serialVersionUID = 1L;
